Expose flow_control paramater from zigpy
Certain boards (such as the Sonoff Zigbee 3.0 USB Dongle Plus P) have a
hardware switch that allows for turning on hardware flow control (given
the correct firmware is flashed). This requires of course that zigpy-cli
also allows for setting this flag to be able to communicate with the
dongle.

Zigpy and zigpy-znp already support hardware flow control fully. No wit
is time that zigpy-cli also offers this option.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>